+++ src/commonMain/kotlin/exchange.dydx.abacus/processor/markets/OrderbookEntryProcessor.kt
@@ -10,7 +10,7 @@ import exchange.dydx.abacus.utils.safeSet
 
 internal class OrderbookEntryProcessor(parser: ParserProtocol) : BaseProcessor(parser) {
     private val orderbookEntryKeyMap = iMapOf(
-        "decimal" to iMapOf(
+        "double" to iMapOf(
             "size" to "size",
             "price" to "price"
         )
